backend/requirements.txt
# FastAPI and web framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
alembic==1.13.1
asyncpg==0.29.0
pgvector==0.2.4

# Pydantic for validation
pydantic==2.5.3
pydantic-settings==2.1.0

# LangChain and AI
langchain==0.1.0
langchain-openai==0.0.2
openai==1.6.1

# Redis for caching
redis==5.0.1

# HTTP client and web scraping
httpx==0.26.0
beautifulsoup4==4.12.2
lxml==4.9.4

# Scheduling
apscheduler==3.10.4

# Security and utilities
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# Testing
pytest==8.0.0
pytest-asyncio==0.23.3
pytest-cov==4.0.0

# Development tools
black==24.0.0
ruff==0.1.9
mypy==1.8.0

# Monitoring and logging
prometheus-client==0.19.0
structlog==23.2.0